Make some things more consistent
