+++ embassy-executor/src/raw/mod.rs
@@ -35,7 +35,6 @@ use core::marker::PhantomData;
 use core::mem;
 use core::pin::Pin;
 use core::ptr::NonNull;
-#[cfg(feature = "drs-scheduler")]
 use core::ptr::addr_of_mut;
 #[cfg(not(feature = "arch-avr"))]
 use core::sync::atomic::AtomicPtr;
@@ -45,16 +44,11 @@ use core::task::{Context, Poll};
 #[cfg(feature = "arch-avr")]
 use portable_atomic::AtomicPtr;
 
-use self::run_queue::RunQueue;
-#[cfg(not(feature = "drs-scheduler"))]
-use self::run_queue::RunQueueItem;
+use self::run_queue::{RunQueue, RunQueueItem};
 use self::state::State;
 use self::util::{SyncUnsafeCell, UninitCell};
 pub use self::waker::task_from_waker;
 use super::SpawnToken;
-
-#[cfg(feature = "drs-scheduler")]
-use cordyceps::{stack, Linked};
 
 #[cfg(feature = "drs-scheduler")]
 pub use run_queue::Deadline;
@@ -98,30 +92,13 @@ pub use run_queue::Deadline;
 /// - 4: A run-queued task exits - `TaskStorage::poll -> Poll::Ready`
 /// - 5: Task is dequeued. The task's future is not polled, because exiting the task replaces its `poll_fn`.
 /// - 6: A task is waken when it is not spawned - `wake_task -> State::run_enqueue`
-#[cfg_attr(feature = "drs-scheduler", repr(C))]
 pub(crate) struct TaskHeader {
-    // TODO(AJM): Make a decision whether we want to support the spicier "pointer recast"/"type punning"
-    // method of implementing the `cordyceps::Linked` trait or not.
-    //
-    // Currently, I do the safer version with `addr_of_mut!`, which doesn't REQUIRE that the first
-    // element is the `links` field, at the potential cost of a little extra pointer math.
-    //
-    // The optimizer *might* (total guess) notice that we are always doing an offset of zero in the
-    // call to `addr_of_mut` in the `impl Linked for TaskHeader` below, and get the best of both worlds,
-    // but right now this is maybe a little over cautious.
-    //
-    // See https://docs.rs/cordyceps/latest/cordyceps/trait.Linked.html#implementing-linkedlinks for
-    // more context on the choices here.
-    #[cfg(feature = "drs-scheduler")]
-    pub(crate) links: stack::Links<TaskHeader>,
-
-    #[cfg(feature = "drs-scheduler")]
-    pub(crate) deadline: SyncUnsafeCell<u64>,
-
-    // TODO(AJM): We could potentially replace RunQueueItem for other runqueue impls, though
-    // right now cordyceps doesn't work on non-atomic systems
-    #[cfg(not(feature = "drs-scheduler"))]
     pub(crate) run_queue_item: RunQueueItem,
+
+    #[cfg(feature = "drs-scheduler")]
+    /// Deadline Rank Scheduler Deadline. This field should not be accessed outside the context of
+    /// the task itself as it being polled by the executor.
+    pub(crate) deadline: SyncUnsafeCell<u64>,
 
     pub(crate) state: State,
     pub(crate) executor: AtomicPtr<SyncExecutor>,
@@ -135,25 +112,6 @@ pub(crate) struct TaskHeader {
     pub(crate) id: u32,
     #[cfg(feature = "trace")]
     all_tasks_next: AtomicPtr<TaskHeader>,
-}
-
-#[cfg(feature = "drs-scheduler")]
-unsafe impl Linked<stack::Links<TaskHeader>> for TaskHeader {
-    type Handle = TaskRef;
-
-    fn into_ptr(r: Self::Handle) -> NonNull<Self> {
-        r.ptr.cast()
-    }
-
-    unsafe fn from_ptr(ptr: NonNull<Self>) -> Self::Handle {
-        let ptr: NonNull<TaskHeader> = ptr;
-        TaskRef { ptr }
-    }
-
-    unsafe fn links(ptr: NonNull<Self>) -> NonNull<stack::Links<TaskHeader>> {
-        let ptr: *mut TaskHeader = ptr.as_ptr();
-        NonNull::new_unchecked(addr_of_mut!((*ptr).links))
-    }
 }
 
 /// This is essentially a `&'static TaskStorage<F>` where the type of the future has been erased.
@@ -242,10 +200,7 @@ impl<F: Future + 'static> TaskStorage<F> {
     pub const fn new() -> Self {
         Self {
             raw: TaskHeader {
-                #[cfg(not(feature = "drs-scheduler"))]
                 run_queue_item: RunQueueItem::new(),
-                #[cfg(feature = "drs-scheduler")]
-                links: stack::Links::new(),
                 #[cfg(feature = "drs-scheduler")]
                 deadline: SyncUnsafeCell::new(0u64),
                 state: State::new(),
